Add files via upload
We're getting some errors when we run comprehensive_school_risk, but hopefully this will get us closer to a Kenya PoC!

Also - ideally, your directory structure should look something like this:

/Users/alice/Developer/UN-Tech-Over
├── data
├── venv
├── distance_analysis
├── ghsl
│   ├── GHS_BUILT_S_GLOBE_R2023A
│   ├── GHS_POP_GLOBE_R2023A
│   └── GHS_SMOD_GLOBE_R2023A
├── main_solution
└── tajikistan_poc

With the tajikistan-specific python in the `tajikistan_poc` directory, and all other python files in the `main_solution` file.  Please read through the code to see if this makes sense - and Ben, please implement this directory structure when you get the chance!